MultiSymplectic Library Functions

MultiSymplectic.Lagrangian_multiplier — Method

Basis for Lagrangian multipliers at the boundary, for 1D.

MultiSymplectic.composite_quadrature — Method

composite_quadrature(breaks::Vector{T}, k::Int) where T<:Real

Generate composite Gauss-Legendre quadrature nodes and weights over multiple intervals defined by breaks.

breaks: strictly increasing vector [x1, x2, ..., xn]

k: order of QuadratureRules

return: (allnodes, allweights)

MultiSymplectic.construct_quadrature_grid — Method

construct_quadrature_grid(R_list::Vector{Int},Num_intervals::Vector{Int})

Generate a grid of quadrature nodes and weights for multiple dimensions using composite Gauss-Legendre quadrature.

R_list: vector specifying the number of quadrature points per interval for each dimension

Num_intervals: vector specifying the number of intervals for each dimension